backend/requirements.txt
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-api>=0.11.0
openai>=1.12.0
httpx>=0.26.0
python-dotenv>=1.0.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
numpy>=1.26.3
aiohttp>=3.9.3
structlog>=24.1.0
PyJWT>=2.8.0
flask>=3.0.0

